src/requirements.txt
adal==1.2.6
applicationinsights==0.11.9
argcomplete==1.12.2
argon2-cffi==21.1.0
async-generator==1.10
attrs==21.2.0
azure-cli-core==2.20.0
azure-cli-telemetry==1.0.6
azure-common==1.1.27
azure-core==1.19.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.1.0
azure-mgmt-resource==12.1.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.9.0
azureml-automl-core==1.34.1
azureml-contrib-dataset==1.34.0
azureml-core==1.34.0
azureml-dataprep==2.22.2
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==1.20.2
azureml-dataset-runtime==1.34.0
azureml-pipeline==1.34.0
azureml-pipeline-core==1.34.0
azureml-pipeline-steps==1.34.0
azureml-sdk==1.34.0
azureml-telemetry==1.34.0
azureml-train==1.34.0
azureml-train-automl-client==1.34.0
azureml-train-core==1.34.0
azureml-train-restclients-hyperdrive==1.34.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
bleach==4.1.0
certifi
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
cloudpickle==1.6.0
colorama==0.4.4
contextlib2==21.6.0
cryptography==3.3
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
distro==1.5.0
docker==5.0.2
dotnetcore2==2.1.20
entrypoints==0.3
fusepy==3.0.1
humanfriendly==9.1
idna==3.2
importlib-metadata==4.8.1
ipykernel
ipython
ipython-genutils
ipywidgets
isodate==0.6.0
jedi==0.17.0
jeepney==0.6.0
Jinja2==3.0.1
jmespath==0.10.0
jsonpickle==2.0.0
jsonschema
jupyter==1.0.0
jupyter-client
jupyter-console
jupyter-core
jupyterlab-pygments
jupyterlab-widgets
kiwisolver
knack==0.8.0rc2
matplotlib
matplotlib-inline
mistune==0.8.4
mmocr
msal==1.10.0
msal-extensions==0.2.2
msrest==0.6.21
msrestazure==0.6.4
nbclient
nbconvert
nbformat
ndg-httpsclient==0.5.1
nest-asyncio
notebook
numpy
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
packaging
pandas==1.1.5
pandocfilters
paramiko==2.7.2
parso
pathspec==0.8.1
pexpect
pickleshare
Pillow
pkginfo==1.7.0
portalocker==1.7.1
prometheus-client
prompt-toolkit
psutil==5.8.0
ptyprocess
pyarrow==1.0.1
pyasn1==0.4.8
pycocotools==2.0.2
pycparser
Pygments
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing
pyrsistent
python-dateutil
pytz==2021.3
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.6
scikit-learn
SecretStorage==3.3.1
Send2Trash
six
tabulate==0.8.9
terminado==0.9.2
testpath
torch==1.8.0
torchvision==0.9.0
tornado
tqdm==4.59.0
traitlets
typing-extensions
urllib3==1.26.6
wcwidth
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.1
zipp

# Celery
celery==5.2.7
